Update xml planet loading for new location of hiring hall information
